Add subdirectory create and delete events for Shared Directory

+++ packages/dds/map/src/directory.ts
@@ -28,6 +28,8 @@ import {
     ISerializedValue,
     ISharedDirectory,
     ISharedDirectoryEvents,
+    ISubDirectoryCreated,
+    ISubDirectoryDeleted,
     IValueChanged,
 } from "./interfaces";
 import {
@@ -1465,6 +1467,11 @@ class SubDirectory extends TypedEventEmitter<IDirectoryEvents> implements IDirec
                     this.serializer,
                     posix.join(this.absolutePath, subdirName)),
             );
+            const event: ISubDirectoryCreated = {
+                key: subdirName,
+                path: this.absolutePath,
+            };
+            this.directory.emit("subDirectoryCreated", event, local, this.directory);
         }
     }
 
@@ -1475,8 +1482,23 @@ class SubDirectory extends TypedEventEmitter<IDirectoryEvents> implements IDirec
      * @param op - The message if from a remote delete, or null if from a local delete
      */
     private deleteSubDirectoryCore(subdirName: string, local: boolean) {
-        // This should make the subdirectory structure unreachable so it can be GC'd and won't appear in snapshots
-        // Might want to consider cleaning out the structure more exhaustively though?
-        return this._subdirectories.delete(subdirName);
+        if (this.hasSubDirectory(subdirName)) {
+            const previousValue = this.getSubDirectory(subdirName);
+            assert(previousValue !== undefined, "Subdirectory should be expected");
+            // This should make the subdirectory structure unreachable so it can be GC'd and won't appear in snapshots
+            // Might want to consider cleaning out the structure more exhaustively though?
+            const successfullyRemoved = this._subdirectories.delete(subdirName);
+            if (successfullyRemoved) {
+                const event: ISubDirectoryDeleted = {
+                    key: subdirName,
+                    path: this.absolutePath,
+                    previousValue,
+                };
+                this.directory.emit("subDirectoryDeleted", event, local, this.directory);
+            }
+            return successfullyRemoved;
+        }
+
+        return false;
     }
 }
